test/requirements.txt
Flask
openai
whisper
python-dotenv
pyaudio
wave
keyboard